AI-Query-Engine-Frontend/requirements.txt
streamlit
requests # if this cause problems you can remove it.
python-dotenv